requirements.txt
Mezzanine==4.2.2
django-modeltranslation==0.12
psycopg2==2.6.1
dj-database-url==0.4.1